notebooks/lab-05-agentcore-observability/requirements.txt
strands-agents>=0.1.0
bedrock-agentcore>=0.1.2
boto3>=1.39.15
